+++ src/test/java/ninja/burpsuite/libs/generic/UIHelperTest.java
@@ -1,5 +1,6 @@
 // Burp Suite Extension Name: Sharpener
-// Released under AGPL see LICENSE for more information
+// Copyright (C) 2021-2026 Soroush Dalili (@irsdl)
+// Released under AGPL v3.0 with additional terms, see the LICENSE and NOTICE files
 // Developed by Soroush Dalili (@irsdl)
 // Project link: https://github.com/irsdl/BurpSuiteSharpenerEx